requirements.txt
opencv-python~=4.9.0.80
numpy~=1.23.5
h11~=0.14.0
jax~=0.4.26
scipy~=1.11.4
Pygments~=2.17.2
rich~=13.7.1
colorama~=0.4.6
tensorflow~=2.12.0
flatbuffers~=23.5.26
requests~=2.31.0
matplotlib~=3.8.2
Pillow~=10.1.0
pip~=23.3.2
distro~=1.9.0
wheel~=0.38.4
Jinja2~=3.1.2
filelock~=3.13.1
rsa~=4.9
pyasn1~=0.5.1
gast~=0.4.0
h5py~=3.10.0
pytz~=2023.3.post1
typing_extensions~=4.8.0
python-dateutil~=2.8.2
ssim~=0.3.0
keras~=2.12.0
PyYAML~=6.0.1
sniffio~=1.3.1
httpcore~=1.0.4
idna~=3.6
certifi~=2023.11.17
six~=1.16.0
pandas~=2.1.3
wrapt~=1.14.1
mdurl~=0.1.2
namex~=0.0.8
setuptools~=65.5.1
patsy~=0.5.6
regex~=2023.12.25
mpmath~=1.3.0
sympy~=1.12
Xplique~=1.3.3
scikit-learn~=1.3.2
torch~=2.1.1
astunparse~=1.6.3
safetensors~=0.4.2
transformers~=4.39.1
tensorboard~=2.12.3
packaging~=23.2
tqdm~=4.66.1
torchvision~=0.16.1
urllib3~=2.1.0
fsspec~=2023.10.0
cachetools~=5.3.2
MarkupSafe~=2.1.3
joblib~=1.3.2
openai~=1.14.2
httpx~=0.27.0
pydantic~=2.6.4
anyio~=4.3.0
pydantic_core~=2.16.3
optree~=0.11.0
plotly~=5.20.0
tenacity~=8.2.3
scikit-image~=0.22.0
chardet~=5.2.0
imageio~=2.33.0
tifffile~=2023.9.26
seaborn~=0.13.1
cycler~=0.12.1
statsmodels~=0.14.2
lazy_loader~=0.3
networkx~=3.2.1
threadpoolctl~=3.2.0
Deprecated~=1.2.14
Markdown~=3.5.2
oauthlib~=3.2.2
Werkzeug~=3.0.1
contourpy~=1.2.0
fonttools~=4.45.1
reportlab~=4.0.8
pyparsing~=3.1.1
termcolor~=2.4.0
kiwisolver~=1.4.5
tokenizers~=0.15.2
pyclustering~=0.10.1.2
SciencePlots~=2.1.1
albumentations~=1.3.1
qudida~=0.0.4
ttach~=0.0.3